+++ scripts/install.sh
@@ -6,10 +6,10 @@ info "start installing xmake ..."
 
 rm -rf build
 
-git submodule update --init
-./configure --verbose
-make
-sudo make install PREFIX=/usr/local
+git submodule update --init || exit 1
+./configure --verbose || exit 1
+make || exit 1
+sudo make install PREFIX=/usr/local || exit 1
 
 ok "install xmake successfully!"